Refactor plugins
Created some new plugins and rearranged existing ones.
