+++ crates/genesis_attributes/src/lib.rs
@@ -518,6 +518,7 @@ impl Plugin for AttributesPlugin {
             .register_component_as::<dyn AttributeDisplay, MaxSize>()
             .register_component_as::<dyn AttributeDisplay, GrowthRate>()
             .register_component_as::<dyn AttributeDisplay, GrabAngle>()
-            .register_component_as::<dyn AttributeDisplay, GrabStrength>();
+            .register_component_as::<dyn AttributeDisplay, GrabStrength>()
+            .init_resource::<Genome>();
     }
 }
