+++ crates/genesis_lib/src/lib.rs
@@ -78,13 +78,11 @@ impl Plugin for GenesisPlugin {
                 SystemStage::parallel().with_system_set(despawn_system_set()),
             )
             .init_resource::<bug_serde::LoadedBlueprint>()
-            .init_resource::<genesis_attributes::Genome>()
+            .init_resource::<statistics::FamilyTree>()
             .insert_resource(config::BACKGROUND)
-            .insert_resource(statistics::FamilyTree::default())
             .insert_resource(spawners)
             .insert_resource(plant_spawn_size)
             .insert_resource(ecosystem)
-            .insert_resource(setup::MindThresholds::new(&config_instance.brain_mutations))
             .add_startup_system_set(setup::setup_system_set())
             .add_system_set(plant_system_set())
             .add_fixed_timestep(Duration::from_secs(10), "family_tree")
